+++ src/roman_datamodels/table_definitions.py
@@ -1,9 +1,8 @@
-import numpy as np
-
-
-DQ_DEF_DTYPE = np.dtype([
-    ("BIT", np.int32),
-    ("VALUE", np.uint32),
-    ("NAME", "S40"),
-    ("DESCRIPTION", "S80")
-])
+# Pleaceholder for future tables, ex:
+#
+# DQ_DEF_DTYPE = np.dtype([
+#     ("BIT", np.int32),
+#     ("VALUE", np.uint32),
+#     ("NAME", "S40"),
+#     ("DESCRIPTION", "S80")
+# ])